pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-google-vertexai>=3.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "mcp-server-calculator>=0.2.0",
    "python-dotenv>=1.1.1",
]
